systemlibs: allow building with PREFIX other than /usr
Some use-cases want to build packages in a different path than /usr. For
example to have a set of packages independent from other system
binaries. This change allows building with bazel build
--define=PREFIX=/some/other/path to search that path instead. The
default of /usr is set in bazelrc so that building with no options will
work as before and setting PREFIX on the commandline or later in the
bazelrc will override that setting if desired. PREFIX is not used by the
bundled build so should not affect that at all.

This also adds a few other standard Make variables that can be
overridden independently if needed.

Signed-off-by: Jason Zaman <jason@perfinion.com>